Implement minmaxnormalization for all pipelines
