+++ clinicadl/clinicadl/train/train_autoencoder.py
