+++ clinicadl/clinicadl/train/train_singleCNN.py
