+++ clinicadl/clinicadl/train/train_singleCNN_bad_data_split.py
